Mudar archivo de constantes a /constants y refactorizar el código para adaptar a los cambios
