+++ lib/bloc/noticia_bloc/noticia_bloc.dart
@@ -103,7 +103,7 @@ import 'package:afranco/bloc/noticia_bloc/noticia_event.dart';
 import 'package:afranco/bloc/noticia_bloc/noticia_state.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:afranco/api/service/noticia_repository.dart';
-import 'package:afranco/constants.dart';
+import 'package:afranco/constants/constants.dart';
 import 'package:watch_it/watch_it.dart';
 
 class NoticiaBloc extends Bloc<NoticiaEvent, NoticiaState> {
